+++ .gitignore
@@ -1,3 +1,3 @@
 .tmp
 result
-setupparser/.direnv
+**/.direnv
